requirements.txt
python-telegram-bot==13.14.*
environs>=9.*
google-cloud-dialogflow>=2.16.*
